+++ NEWS.md
@@ -1,6 +1,12 @@
 Release Notes
 =============
 
+Version 0.1.2 (2023-01-10)
+--------------------------
+### Bug fixes
+* Fixed a bug that forces the electrolyser to operate when the parameter Minimum_load was not equal to 0
+* Fixed a bug, that required stack replacement in all years after the initial stack replacement 
+
 Version 0.1.1 (2022-12-12)
 --------------------------
 Updated Readme